clean up and consolidate code
Removed a lot of unused variables and imports.
Consolidated many methods such as setbox1value, setbox2value, etc. and
play1, play2, play3, etc. into just 2 methods setboxNvalue and playN.
Moved a number of hard-coded values into variables to make it easier to
adjust: grid size, midi-note values, kinect distance for detection.

All this is un-tested...

One thing of note which should definitely be tested, there were 2
variables which were created and appeared to have been meant for the
same thing (previnbox[] and pinbox[]). In many places in checked a for a
previnbox value and in one place it checked for a pinbox value. It never
set any previnbox values though it did set pinbox values.  My guess is
that this was meant to prevent multiple triggers while standing in one
place, though what we saw was multiple triggers. Obviously if it never
set any values to check it wouldn't work.... Anyway, I combined the two
variables into one. I don't know if this will actually take care of the
repeated trigger effect or will possibly cause some other unintended
effect since I didn't have time to work through the logic and figure it
out... maybe it'll work or maybe we need to just get rid of the
previnbox variable as it is and re-do it.